IoSync handles concurrent loading and saving of cache entries
It can be used directly as a RemovalListener and CacheLoader.
The onRemoval() method enqueues values for writing and returns immediately.
IoSync takes care of directly returning values that reloaded are while
they are written. It also ensures that the final state of a value that is
enqueued for removal several times is the state that is eventually written.